pyproject.toml
[project]
name = "azure-openai-workshop"
version = "0.1.0"
description = "Azure OpenAI Workshop - Learn to build AI applications with Azure AI Foundry"
requires-python = ">=3.12"
dependencies = [
    # Core dependencies for Azure AI and OpenAI
    "python-dotenv>=1.0.0",
    "openai>=1.58.1",
    "azure-ai-projects>=1.0.0b1",
    "azure-ai-inference>=1.0.0b1",
    "azure-identity>=1.15.0",
    "azure-cognitiveservices-speech>=1.37.0",
    
    # Data manipulation and analysis
    "pandas>=2.2.1",
    "numpy>=1.26.4",
    
    # Visualization and image processing
    "matplotlib>=3.8.3",
    "pillow>=10.2.0",
    
    # Text processing and tokenization
    "tiktoken>=0.6.0",
    
    # Azure Search (if needed for advanced examples)
    "azure-search-documents>=11.6.0b1",
    "azure-storage-blob>=12.19.0",
    
    # Document processing
    "markitdown>=0.0.1a2",
    
    # Jupyter notebook support
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.8.0",
]

# Tracing and observability dependencies (for workshop 2)
tracing = [
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-openai>=0.22.0",
    "azure-monitor-opentelemetry>=1.2.0",
]

# Agent framework dependencies (for workshop 3)
agents = [
    "semantic-kernel>=1.0.0",
    # "autogen-agentchat>=0.2.0",  # Temporarily disabled due to dependency conflicts
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/azure_openai_workshop"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.8.0",
]
